+++ gulpfile.js
@@ -59,6 +59,7 @@ gulp.task('script', () => {
 		external: [
 			'child_process',
 			'fs',
+			'jsonc-parser',
 			'path',
 			'vscode',
 		],
@@ -81,6 +82,7 @@ gulp.task('script', () => {
 			globals: {
 				child_process: 'child_process',
 				fs: 'fs',
+				jsoncParser: 'jsonc-parser',
 				path: 'path',
 				vscode: 'vscode',
 			},
@@ -90,7 +92,7 @@ gulp.task('script', () => {
 	
 });
 
-gulp.task('build', gulp.series('clean', 'script'));
+gulp.task('build', gulp.series('clean', 'icons:fix', 'script'));
 
 gulp.task('watch', () => {