+++ src/statusbar/colors.ts
@@ -1,0 +1,59 @@
+//	Imports ____________________________________________________________________
+
+import { StatusbarColors } from '../@types/workspaces';
+
+//	Variables __________________________________________________________________
+
+
+
+//	Initialize _________________________________________________________________
+
+
+
+//	Exports ____________________________________________________________________
+
+export const colors:StatusbarColors[] = [
+	{
+		'statusBar.background': null,
+		'statusBar.foreground': null,
+		'statusBarItem.hoverBackground': null,
+	},
+	{ // purple
+		'statusBar.background': '#691d7c',
+		'statusBar.foreground': '#ffffff',
+		'statusBarItem.hoverBackground': '#7a3c8a',
+	},
+	{ // blue
+		'statusBar.background': '#0178cf',
+		'statusBar.foreground': '#ffffff',
+		'statusBarItem.hoverBackground': '#1f8ad2',
+	},
+	{ // green
+		'statusBar.background': '#28632b',
+		'statusBar.foreground': '#ffffff',
+		'statusBarItem.hoverBackground': '#327e36',
+	},
+	{ // yellow
+		'statusBar.background': '#a79c21',
+		'statusBar.foreground': '#ffffff',
+		'statusBarItem.hoverBackground': '#beb012',
+	},
+	{ // orange
+		'statusBar.background': '#ce7e1c',
+		'statusBar.foreground': '#ffffff',
+		'statusBarItem.hoverBackground': '#ff8e00',
+	},
+	{ // red
+		'statusBar.background': '#d63f26',
+		'statusBar.foreground': '#ffffff',
+		'statusBarItem.hoverBackground': '#e65b42',
+	},
+	{ // grey
+		'statusBar.background': '#7d7d7d',
+		'statusBar.foreground': '#ffffff',
+		'statusBarItem.hoverBackground': '#8c8c8c',
+	},
+];
+
+//	Functions __________________________________________________________________
+